For iNES MapperInfo, use uint16 instead of uint8 for 2.0 compatibility

+++ src/Vs/mapper000.cpp
@@ -28,7 +28,7 @@ void	MAPINT	Unload (void)
 	VS::Unload();
 }
 
-uint8 MapperNum = 0;
+uint16 MapperNum = 0;
 } // namespace
 
 const MapperInfo MapperInfo_000 =
