+++ src/Vs/mapper004.cpp
@@ -48,7 +48,7 @@ void	MAPINT	Unload (void)
 	VS::Unload();
 }
 
-uint8 MapperNum = 4;
+uint16 MapperNum = 4;
 } // namespace
 
 const MapperInfo MapperInfo_004 =
