+++ src/iNES/mapper004.cpp
@@ -43,7 +43,7 @@ void	MAPINT	Unload (void)
 	MMC3::Unload();
 }
 
-uint8 MapperNum = 4;
+uint16 MapperNum = 4;
 } // namespace
 
 const MapperInfo MapperInfo_004 =
