+++ src/iNES/mapper007.cpp
@@ -33,7 +33,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 7;
+uint16 MapperNum = 7;
 } // namespace
 
 const MapperInfo MapperInfo_007 =
