+++ src/iNES/mapper009.cpp
@@ -32,7 +32,7 @@ void	MAPINT	Unload (void)
 	MMC2::Unload();
 }
 
-uint8 MapperNum = 9;
+uint16 MapperNum = 9;
 } // namespace
 
 const MapperInfo MapperInfo_009 =
