+++ src/iNES/mapper031.cpp
@@ -53,7 +53,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 31;
+uint16 MapperNum = 31;
 } // namespace
 
 const MapperInfo MapperInfo_031 =
