+++ src/iNES/mapper033.cpp
@@ -83,7 +83,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 33;
+uint16 MapperNum = 33;
 } // namespace
 
 const MapperInfo MapperInfo_033 =
