+++ src/iNES/mapper040.cpp
@@ -82,7 +82,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 40;
+uint16 MapperNum = 40;
 } // namespace
 
 const MapperInfo MapperInfo_040 =
