+++ src/iNES/mapper041.cpp
@@ -56,7 +56,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 41;
+uint16 MapperNum = 41;
 } // namespace
 
 const MapperInfo MapperInfo_041 =
