+++ src/iNES/mapper046.cpp
@@ -52,7 +52,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 46;
+uint16 MapperNum = 46;
 } // namespace
 
 const MapperInfo MapperInfo_046 =
