+++ src/iNES/mapper048.cpp
@@ -127,7 +127,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 48;
+uint16 MapperNum = 48;
 } // namespace
 
 const MapperInfo MapperInfo_048 =
