+++ src/iNES/mapper050.cpp
@@ -85,7 +85,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 50;
+uint16 MapperNum = 50;
 } // namespace
 
 const MapperInfo MapperInfo_050 =
