+++ src/iNES/mapper051.cpp
@@ -65,7 +65,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 51;
+uint16 MapperNum = 51;
 } // namespace
 
 const MapperInfo MapperInfo_051 =
