+++ src/iNES/mapper053.cpp
@@ -63,7 +63,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 53;
+uint16 MapperNum = 53;
 } // namespace
 
 const MapperInfo MapperInfo_053 =
