+++ src/iNES/mapper058.cpp
@@ -32,7 +32,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 58;
+uint16 MapperNum = 58;
 } // namespace
 
 const MapperInfo MapperInfo_058 =
