+++ src/iNES/mapper072.cpp
@@ -57,7 +57,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 72;
+uint16 MapperNum = 72;
 } // namespace
 
 const MapperInfo MapperInfo_072 =
