+++ src/iNES/mapper075.cpp
@@ -92,7 +92,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 75;
+uint16 MapperNum = 75;
 } // namespace
 
 const MapperInfo MapperInfo_075 =
