+++ src/iNES/mapper076.cpp
@@ -89,7 +89,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 76;
+uint16 MapperNum = 76;
 } // namespace
 
 const MapperInfo MapperInfo_076 =
