+++ src/iNES/mapper087.cpp
@@ -45,7 +45,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 87;
+uint16 MapperNum = 87;
 } // namespace
 
 const MapperInfo MapperInfo_087 =
