+++ src/iNES/mapper119.cpp
@@ -39,7 +39,7 @@ void	MAPINT	Unload (void)
 	MMC3::Unload();
 }
 
-uint8 MapperNum = 119;
+uint16 MapperNum = 119;
 } // namespace
 
 const MapperInfo MapperInfo_119 =
