+++ src/iNES/mapper125.cpp
@@ -48,7 +48,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 125;
+uint16 MapperNum = 125;
 } // namespace
 
 const MapperInfo MapperInfo_125 =
