+++ src/iNES/mapper127.cpp
@@ -100,7 +100,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 127;
+uint16 MapperNum = 127;
 } // namespace
 
 const MapperInfo MapperInfo_127 =
