+++ src/iNES/mapper138.cpp
@@ -91,7 +91,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 138;
+uint16 MapperNum = 138;
 } // namespace
 
 const MapperInfo MapperInfo_138 =
