+++ src/iNES/mapper140.cpp
@@ -44,7 +44,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 140;
+uint16 MapperNum = 140;
 } // namespace
 
 const MapperInfo MapperInfo_140 =
