+++ src/iNES/mapper145.cpp
@@ -48,7 +48,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 145;
+uint16 MapperNum = 145;
 } // namespace
 
 const MapperInfo MapperInfo_145 =
