+++ src/iNES/mapper146.cpp
@@ -48,7 +48,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 146;
+uint16 MapperNum = 146;
 } // namespace
 
 const MapperInfo MapperInfo_146 =
