+++ src/iNES/mapper147.cpp
@@ -50,7 +50,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 147;
+uint16 MapperNum = 147;
 } // namespace
 
 const MapperInfo MapperInfo_147 =
