+++ src/iNES/mapper148.cpp
@@ -31,7 +31,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 148;
+uint16 MapperNum = 148;
 } // namespace
 
 const MapperInfo MapperInfo_148 =
