+++ src/iNES/mapper150.cpp
@@ -48,7 +48,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 150;
+uint16 MapperNum = 150;
 } // namespace
 
 const MapperInfo MapperInfo_150 =
